Update Motd
At the moment, there is a gamebreaking bug involving the movement of the player in 1.0.24 and above. Please refrain downloading this version until further notice. If you have already downloaded said version(s) then please attempt to downgrade to 1.0.23.4. This update board
will change when this glitch has been patched.